--- src/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Fri Jan  5 08:30:14 2024
-
-@author: Anita Karsa, University of Cambridge, UK
-"""
