+++ unfolding/_utils.py
@@ -1,0 +1,14 @@
+import numpy as np
+
+
+def dummy():
+    """Create a dummy image and label"""
+    x, y, z = np.meshgrid(
+        np.linspace(0, 1, 100), np.linspace(0, 1, 100), np.linspace(0, 1, 100)
+    )
+
+    image = x + y + z
+
+    label = ((x - 0.5) ** 2 + (y - 0.5) ** 2 + (z - 0.5) ** 2) < 0.2**2
+
+    return image, label
